refactor: Integrate ActionType.hpp and ActionEnums.hpp
